+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Public api is not require adding source name for getting default image
 * Moved docker image from alpine to distro-less
 * Implemented app logger with saving to file.
+* Fixed some warnings and better error handling.
 
 # 0.1.2 (07.01.2022)